pyproject.toml
[project]
name = "image-segmentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fiftyone>=0.25.0",
    "numpy>=2.0.0",
    "pillow>=12.0.0",
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.13",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]
